requirements.txt
django
gunicorn
google-cloud-storage
